requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
boltons==20.2.1
chardet==4.0.0
fire==0.4.0
future==0.18.2
idna==3.1
importlib-metadata==3.7.3
multidict==5.1.0
pathutils==1.0
pefile==2019.4.18
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
six==1.15.0
soupsieve==2.2.1
termcolor==1.1.0
typing-extensions==3.7.4.3
yarl==1.6.3
zipp==3.4.1
